chức năng đặt hằng

+++ database/migrations/2022_09_24_093641_create_customers_table.php
@@ -20,7 +20,7 @@ return new class extends Migration
             $table->string('email', 50);
             $table->string('address', 150);
             $table->text('note');
-            $table->decimal('payment');
+            $table->bigInteger('payment');
             $table->timestamps();
         });
     }
